requirements.txt
pycrypto
sqlalchemy
urllib3